Split api.rs into api/ directory by domain
Refs #107.

api.rs (1489 lines) → api/ (9 files):

  mod.rs         PadzApi<S> struct, new(), paths(), re-exports
                 (TodoStatus, PadFilter, PadStatusFilter, CmdMessage,
                 MessageLevel, CmdResult, PadUpdate, PadzPaths)
  selectors.rs   parse_selectors*, normalize_*, looks_like_short_uuid,
                 canonicalize_or_self
  crud.rs        create_pad, get_pads, view_pads, delete_pads,
                 delete_completed_pads, update_pads,
                 update_pads_from_content, restore_pads, purge_pads,
                 archive_pads, unarchive_pads
  status.rs      pin_pads, unpin_pads, complete_pads, reopen_pads,
                 move_pads, propagate_status
  transfer.rs    export_pads*, import_pads, transfer_pads_to/from
  tags.rs        list_tags, list_pad_tags, create_tag, delete_tag,
                 rename_tag, add_tags_to_pads, remove_tags_from_pads
  init.rs        init, init_link, init_unlink
  util.rs        doctor, pad_paths, pad_uuids, get_path_by_id,
                 refresh_pad, remove_pad
  format.rs      impl PadzApi<FileStore>::create_pad_with_format
  test_support.rs (cfg(test)) shared make_api()/make_store() helpers

Public API is unchanged. `PadzApi::<method>` still resolves through the
re-aggregated impl blocks; `padzapp::api::{PadzApi, PadFilter,
PadStatusFilter, TodoStatus, CmdMessage, MessageLevel, PadzPaths}`
remain at their previous paths.

Tests travel with their subject matter:
  - parse_selectors / normalize_* → selectors.rs
  - CRUD facade tests → crud.rs
  - pin/move tests → status.rs
  - import test → transfer.rs
  - tag tests → tags.rs
  - doctor/paths/refresh/remove → util.rs
  - editor-flow regression → mod.rs
